+++ src/frontend/app/store/types/recently-visited.types.ts
@@ -4,6 +4,16 @@ export interface IEntityHit {
   guid: string;
   date: number;
 }
+export interface IRecentlyVisitedEntity extends IFavoriteTypeInfo {
+  guid: string;
+  name: string;
+  entityId: string;
+  prettyType: string;
+  prettyEndpointType: string;
+  endpointId: string;
+  routerLink?: string;
+}
+
 export interface IRecentlyVisitedEntities {
   [guid: string]: IRecentlyVisitedEntity;
 }
@@ -14,13 +24,4 @@ export interface IRecentlyVisitedState {
 
 export interface IRecentlyVisitedEntityDated extends IRecentlyVisitedEntity {
   date: number;
-}
-export interface IRecentlyVisitedEntity extends IFavoriteTypeInfo {
-  guid: string;
-  name: string;
-  entityId: string;
-  prettyType: string;
-  prettyEndpointType: string;
-  endpointId: string;
-  routerLink?: string;
 }
